+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/JsonValue.java
@@ -5,14 +5,19 @@ import org.apache.commons.lang3.builder.HashCodeBuilder;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.DoubleNode;
+import com.fasterxml.jackson.databind.node.LongNode;
 
 import java.math.BigInteger;
 import java.time.Instant;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 public class JsonValue extends Value {
 
-    private final JsonNode value;
+    private JsonNode value;
 
     private JsonValue(JsonNode value) {
         this.value = value;
@@ -25,7 +30,7 @@ public class JsonValue extends Value {
             checkInt8(val);
             return (byte) val;
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to int8.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to int8, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -35,7 +40,7 @@ public class JsonValue extends Value {
             checkInt8U(val);
             return (short) val;
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to int8U.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to int8U, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -45,7 +50,7 @@ public class JsonValue extends Value {
             checkInt16(val);
             return (short) val;
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to int16.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to int16, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -55,7 +60,7 @@ public class JsonValue extends Value {
             checkInt16U(val);
             return (int) val;
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to int16U.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to int16U, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -65,7 +70,7 @@ public class JsonValue extends Value {
             checkInt32(val);
             return (int) val;
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to int32.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to int32, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -75,7 +80,7 @@ public class JsonValue extends Value {
             checkInt32U(val);
             return val;
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to int32U.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to int32U, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -85,7 +90,7 @@ public class JsonValue extends Value {
             checkInt64(val);
             return val.longValue();
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to int64.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to int64, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -95,17 +100,17 @@ public class JsonValue extends Value {
             checkInt64U(val);
             return val;
         } 
-        throw new UnsupportedOperationException("Cannot convert from JSON to int64U.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to int64U, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
     public float getFloat32() {
         if (value.isNumber()) {
-            double val = value.doubleValue();
+            float val = value.floatValue();
             checkFloat32(val);
-            return (float) val;
+            return val;
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to float32.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to float32, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -115,7 +120,7 @@ public class JsonValue extends Value {
             checkFloat32(val);
             return val;
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to float64.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to float64, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -134,21 +139,35 @@ public class JsonValue extends Value {
             return value.booleanValue();
         } else if (value.isNumber()) {
             return value.longValue() != 0;
+        } else if (value.isTextual()) {
+            return value.asBoolean();
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON to boolean.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to boolean, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
     public EnumRecord getEnum() {
         if (value.isTextual()) {
             return new EnumRecord(value.textValue(), null, null);
+        } else if (value.isIntegralNumber()) {
+            return new EnumRecord(value.asText(), null, null);
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON object to enum.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON object to enum, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
     public Map<String, Boolean> getBitmap() {
-        throw new UnsupportedOperationException("Cannot convert from JSON to bitmap.");
+        if (value.isObject()) {
+            Map<String, Boolean> bm = new HashMap<>();
+            // iterate through children and convert to boolean
+            value.fields().forEachRemaining(e -> {
+                if (e.getValue().isValueNode()) {
+                    bm.put(e.getKey(), e.getValue().asBoolean());
+                }
+            });
+            return bm;
+        }
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON to bitmap, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -158,7 +177,7 @@ public class JsonValue extends Value {
         } else if (value.isNumber()) {
             return Instant.ofEpochMilli(value.longValue());
         }
-        throw new UnsupportedOperationException("Cannot convert from JSON object to datetime.");
+        throw new UnsupportedOperationException(String.format("Cannot convert from JSON object to datetime, node type is '%s'.", value.getNodeType()));
     }
 
     @Override
@@ -168,18 +187,29 @@ public class JsonValue extends Value {
 
     @Override
     public JsonValue[] asArray() {
-        // cannot create actual JSON array here
+        if (value.isArray()) {
+            // make array of JsonValue from each element
+            final List<JsonValue> arr = new ArrayList<>();
+            value.elements().forEachRemaining(e -> arr.add(JsonValue.of(e)));
+            return arr.toArray(new JsonValue[0]);
+        }
         return new JsonValue[]{this};
     }
 
     @Override
     public void absValue() {
-        // not required
+        if (value.isFloatingPointNumber()) {
+            value = DoubleNode.valueOf(Math.abs(value.doubleValue()));
+        } else if (value.isIntegralNumber()) {
+            value = LongNode.valueOf(Math.abs(value.longValue()));
+        }
     }
 
     @Override
     public void roundToInt() {
-        // not required
+        if (value.isFloatingPointNumber()) {
+            value = LongNode.valueOf(Math.round(value.doubleValue()));
+        }
     }
 
     @Override